Csv and support information screen
App can open a csv with potential help for the victim and based on selection made in a drop down menu, can look for the relevant help and display it. Need to add some buttons to scroll through the different helper and add an option to let the user provide the csv.