Finished merging my code with master
Most of my code is inside the main activity, in "getHelp" a function linked to the "get help" button on the help fragment. Pressing it will activate that function which reads the spinner, open the csv, finds the appropiate rows to show and displays them with links.
Main activity has 500loc and code is a bit messy / hard to read. Maybe we should clean it up a bit?